feat(frontend): Add 'insert' route for adding a product to the database
The 'insert' route has been implemented in the frontend to allow the insertion of products into the database. Additionally, a product data input form has been created to enable users to complete the necessary information. The JSON format that will be sent is available for review.

Furthermore, a global state has been established in the 'insert' component to avoid prop drilling, improving data management in the application.